Stripped out the LTI and restructured for api keys.

+++ routes/blueprints.js
@@ -5,75 +5,46 @@ const BLUEPRINT = require('../models/blueprint');
 const SUBMETRIC = require('../models/submetric');
 const SUBMETRIC_TYPES = require('../metric-types');
 const DEBUG = require('debug')('eval:routing');
-const PASSPORT = require('passport');
-const AUTH = require('../includes/authentication')
 
-var router = EXPRESS.Router();
+var router = EXPRESS.Router({ mergeParams: true });
 
-router.use( AUTH.require_login );
-
-/* GET home page. */
-router.all('/', function(req, res, next) {
+router.get( '/list/:api_key', function( req, res ) {
+	// TODO: Filter by api key
 	BLUEPRINT.findAll().then( function( results ) {
-		res.status(200).render( 'blueprints/list', {
-			title: "Rubrics List",
-			path: req.originalUrl,
-			blueprints: results,
-		} );
+		res.status(200).send(results);
 	} );
-});
+} );
 
-router.get('/create', function( req, res ) {
-	var blueprint = { options: {} };
+router.get('/edit/:transaction_id', function( req, res ) {
+	if ( UTIL.is_missing_attributes( ['blueprint_id'], req.params.transaction, res ) ) { return; }
+	var blueprint_id = req.params.transactions.blueprint_id || null;
+	var promises = [];
+
+	if ( blueprint_id ) {
+		promises.push( BLUEPRINT.findOne( {
+			where: { blueprint_id: blueprint_id },
+			include: [{ model: SUBMETRIC }],
+		} ) );
+	}
+
 	var metric_types = [];
-
 	for ( var slug in SUBMETRIC_TYPES ) {
 		if ( SUBMETRIC_TYPES[slug].valid_as_submetric !== false ) {
 			metric_types.push( SUBMETRIC_TYPES[slug] );
 		}
 	}
 
-	res.status(200).render( 'blueprints/editor', {
-		title: "Create Rubric",
-		path: req.originalUrl,
-		blueprint: blueprint,
-		submetrics: [],
-		metric_types: metric_types,
-	} );
-});
-
-router.get('/edit/:blueprint_id', function( req, res ) {
-	var promises = [];
-
-	promises.push( BLUEPRINT.findOne( {
-		where: { blueprint_id: req.params.blueprint_id },
-		include: [{ model: SUBMETRIC }],
-	} ) );
-
-	PROMISE.all( promises ).spread( function( blueprint, submetrics ) {
-		if ( blueprint == null ) {
-			res.status(404).send( "No rubric blueprint #" + req.params.blueprint_id + " found." );
-			return;
-		}
-
-		var metric_types = [];
-		for ( var slug in SUBMETRIC_TYPES ) {
-			if ( SUBMETRIC_TYPES[slug].valid_as_submetric !== false ) {
-				metric_types.push( SUBMETRIC_TYPES[slug] );
-			}
-		}
-
+	PROMISE.all( promises ).spread( function( blueprint ) {
 		res.status(200).render( 'blueprints/editor', {
-			title: "Edit Rubric",
-			path: req.originalUrl,
-			blueprint: blueprint,
+			title: blueprint != null ? "Edit Rubric" : "Create Rubric",
+			blueprint: blueprint != null ? blueprint : { options: {} },
 			metric_types: metric_types,
 		} );
 	} );
 });
 
-function save_blueprint( req, res, next ) {
-	var blueprint_id = req.params.blueprint_id || null;
+router.post( '/save/:transaction_id', function( req, res, next ) {
+	var blueprint_id = req.params.transaction.blueprint_id || null;
 	var data = req.body;
 
 	var promise = null;
@@ -81,7 +52,7 @@ function save_blueprint( req, res, next ) {
 	delete data.submetrics;
 
 	if ( blueprint_id == null ) {
-		DEBUG( "Saving blueprint", data );
+		DEBUG( "Creating blueprint", data );
 		promise = BLUEPRINT.create( data );
 	} else {
 		DEBUG( "Updating blueprint", blueprint_id, data );
@@ -92,7 +63,7 @@ function save_blueprint( req, res, next ) {
 
 	promise.then( function( blueprint ) {
 		blueprint_id = blueprint_id || blueprint.blueprint_id;
-		DEBUG( "Blueprint stored", blueprint_id, "with", submetrics.length, "submetrics" );
+		DEBUG( "Blueprint saved", blueprint_id, "with", submetrics.length, "submetrics" );
 
 		var submetric_ids = [];
 
@@ -128,36 +99,15 @@ function save_blueprint( req, res, next ) {
 			} );
 		} );
 	} );
-}
-
-router.post( '/save/:blueprint_id', save_blueprint );
-router.post( '/save', save_blueprint );
-
-router.get( '/destroy/:blueprint_id', function( req, res ) {
-	var blueprint_id = req.params.blueprint_id;
-
-	BLUEPRINT.findById( blueprint_id ).then( function( blueprint ) {
-		if ( blueprint != null ) {
-			res.status(200).render( 'blueprints/destroy', {
-				title: "Delete Rubric",
-				blueprint: blueprint,
-			} );
-		} else {
-			res.status(404).render('error', {
-				message: "There is no Rubric blueprint with id #" + blueprint_id,
-				error: {
-					status: "404 Blueprint Not Found",
-				},
-			});
-		}
-	} );
 } );
 
-router.post( '/destroy/:blueprint_id', function( req, res ) {
+router.post( '/destroy/:transaction_id', function( req, res ) {
+	if ( UTIL.is_missing_attributes( ['blueprint_id'], req.params.transaction, res ) ) { return; }
+	
 	BLUEPRINT.destroy( {
-		where: { blueprint_id: req.params.blueprint_id },
+		where: { blueprint_id: req.params.transaction.blueprint_id },
 	} ).then( function() {
-		res.status(303).redirect( '/blueprints' );
+		res.status(200).end();
 	} );
 } );
